Reimplemented missing "Components" as Nodes
- Converted Sprite components, animator and Editor camera controller.
- Hopefully the last commit required before converting test project to use the node system. If everything works as it did before, performance testing will follow. If this turns out to be more performant / stable, then it will be merged into the development branch. After further testing and refining, it should be pushed to main.

+++ ChaosEngine/src/Chaos/Nodes/Node.cpp
@@ -10,7 +10,7 @@ namespace Chaos
 	{
 		Level* level = Level::Get();
 		level->Nodes[level->NodeCount][0] = this;
-		ID = level->NodeCount;
+		ID = (uint32_t)level->NodeCount;
 		level->NodeCount++;
 	}
